+++ ChangeLog.md
@@ -3,6 +3,11 @@ Unittests change log
 
 ## ?.?.? / ????-??-??
 
+* Changed `TestSuite::addTestClass()` to return class added instead of
+  list of ignored test methods. The latter is dependant on the internal
+  implementation and also not used anywhere.
+  (@thekid)
+
 ## 6.5.0 / 2015-09-27
 
 * **Heads up: Bumped minimum PHP version required to PHP 5.5**. See PR #4